utils/config.py
"""
Configuration settings for the project.
"""

import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_MODEL = os.getenv("OPENAI_MODEL", "gpt-4o-mini")
OPENAI_TEMPERATURE = float(os.getenv("OPENAI_TEMPERATURE", "0.7"))
OPENAI_MAX_TOKENS = int(os.getenv("OPENAI_MAX_TOKENS", "1500"))
REQUEST_TIMEOUT = int(os.getenv("REQUEST_TIMEOUT", "10"))
CACHE_ENABLED = os.getenv("CACHE_ENABLED", "False").lower() == "true"
CACHE_DIR = os.getenv("CACHE_DIR", ".cache")
