utils/requirements.txt
openai
requests
beautifulsoup4
python-dotenv